requirements.txt
Flask==0.12.2
Flask-Script==2.0.5
python-decouple==3.1
imgurpython==1.1.7
celery==4.1.0


